% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram_network.R
\name{bigram_network}
\alias{bigram_network}
\title{Generate bigram network}
\arguments{
\item{data}{the sentiment140 train or test data containing variables \code{user} for username, \code{date} for date and
\code{text} for text of the tweet.}

\item{user_list}{a vector of users for which to filter the dataset}

\item{start_date_time}{input start_date_time in POSIXct format on which to filter the dataset}

\item{end_date_time}{input end_date_time in POSIXct format on which to filter the dataset}

\item{keyword_list}{a list of string keywords on which to filter the dataset}

\item{counts_quantile}{the quantile beyond which to visualize the bigrams}
}
\value{
a list object with \code{raw} filtered dataframe, \code{bigram_counts} aggregated dataframe that holds the frequency counts of bigrams and a \code{plot} representing the network.
}
\description{
Generate bigram network to analyze given data with filters for all most common bigrams excluding stop words.
}
\examples{
bigram_network()

}
\keyword{analysis}
\keyword{visualization}
