% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\docType{data}
\name{config_params}
\alias{config_params}
\title{Constants for the package}
\format{An object of class \code{list} of length 3.}
\usage{
config_params
}
\description{
Constants for the package
}
\section{Fields}{

\describe{
\item{\code{max_input_data_size}}{max # of rows in the data frame, to control the response time}
}}

\keyword{datasets}
