% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cor_network.R
\name{cor_words}
\alias{cor_words}
\title{Pairwise correlation of words in given dataset}
\usage{
cor_words(data)
}
\arguments{
\item{data}{dataframe containing text}
}
\value{
aggregated dataframe including words and their pairwise correlations
}
\description{
Tokenize text and return dataframe with pairwise word correlation per user for most commonly occuring words
}
\keyword{bigrams}
