% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram_network.R
\name{count_bigrams}
\alias{count_bigrams}
\title{Count bigrams in given dataset}
\usage{
count_bigrams(data)
}
\arguments{
\item{data}{dataframe containing text}
}
\value{
aggregated dataframe including bigrams and their counts
}
\description{
Count bigrams in the given dataset while excluding stop words
}
\keyword{bigrams}
