% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\name{create_conv1d_model}
\alias{create_conv1d_model}
\title{Create 1-Dimensional Convolutional Network model object}
\usage{
create_conv1d_model(max_words, embedding_dim, maxlen, conv1d_filters,
  conv1d_kernel_size, conv1d_pool_size)
}
\arguments{
\item{max_words}{Maximum number of words to consider using word frequency measure.}

\item{embedding_dim}{Output dimension of the embedding layer.}

\item{maxlen}{Maximum length of a sequence.}

\item{conv1d_filters}{Number of filters i.e. output dimension for convolution layers.}

\item{conv1d_kernel_size}{Window size for convolution layers.}

\item{conv1d_pool_size}{Pool size for max pooling.}
}
\value{
1-Dimensional Convolutional Network model object
}
\description{
Create 1-Dimensional Convolutional Network model object
}
