% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\name{create_lstm_model}
\alias{create_lstm_model}
\title{Create LSTM model object}
\usage{
create_lstm_model(max_words, embedding_dim, maxlen, lstm_units)
}
\arguments{
\item{max_words}{Maximum number of words to consider using word frequency measure.}

\item{embedding_dim}{Output dimension of the embedding layer.}

\item{maxlen}{Maximum length of a sequence.}

\item{lstm_units}{Number of units i.e. output dimension of lstm layer.}
}
\value{
LSTM model object
}
\description{
Create LSTM model object
}
