% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_freq_by_polarity.R
\name{freq_by_polarity}
\alias{freq_by_polarity}
\title{Query the dataset for frequency by polarity based on several filters}
\arguments{
\item{data}{the sentiment dataset containing variables \code{user} for username, \code{date} for date and
\code{text} for text of the tweet. default sentiment140_train dataset}

\item{user_list}{a vector of users for which to filter the dataset.}

\item{start_date_time}{input start_date_time in POSIXct format on which to filter the dataset}

\item{end_date_time}{input end_date_time in POSIXct format on which to filter the dataset}

\item{keyword_list}{a list of string keywords on which to filter the dataset}
}
\value{
a list object with \code{raw} filtered dataframe, \code{frequency} dataframe that holds the frequency counts by polarity and a \code{plot} depicting the relationship between the two
}
\description{
Query the dataset for frequency by polarity based on several filters and return the raw data, frequency data and a \code{plot} for the same.
}
\examples{
library(lubridate)
sample_data <- data.frame("user" = c("test_user1", "test_user1", "test_user2"),
                          "date" = c(lubridate::as_datetime("2009-04-01"),
                                     lubridate::as_datetime("2009-04-05"),
                                     lubridate::as_datetime("2009-04-10")),
                          "text" = c("Sample tweet 1 from user1",
                                     "Sample tweet 2 from user1",
                                     "Sample tweet 1 from user2"),
                         "polarity" = c(0, 4, 4),
                         "id" = c(1,2,3))
freq_by_polarity(data = sample_data,
                 user_list = c("", "test_user1"),
                 start_date_time = lubridate::as_datetime("2009-03-30"),
                 end_date_time = lubridate::as_datetime("2009-06-30"),
                 keyword_list = c("tweet 2"))

}
\keyword{query}
\keyword{visualization}
