% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\name{generate_embedding_matrix}
\alias{generate_embedding_matrix}
\title{Generate Embedding Matrix for given word index}
\usage{
generate_embedding_matrix(word_index, embedding_dim, max_words,
  glove_file_path)
}
\arguments{
\item{word_index}{word index as generated by the tokenizer.}

\item{embedding_dim}{Output dimension of the embedding layer.}

\item{max_words}{Maximum number of words to consider using word frequency measure.}

\item{glove_file_path}{File path location for glove embeddings.}
}
\value{
Embedding Matrix object
}
\description{
Generate Embedding Matrix for given word index
}
