% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\name{generate_training_data}
\alias{generate_training_data}
\title{Generate training data}
\usage{
generate_training_data(data, sequences, maxlen, seed)
}
\arguments{
\item{data}{Input data.}

\item{sequences}{sequences generated by using tokenizer on text}

\item{maxlen}{Maximum length of a sequence.}

\item{seed}{Seed for shuffling data}
}
\value{
List object with \code{x_train} and \code{y_train} objects.
}
\description{
Generate training data
}
