% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{get_keyword_condition}
\alias{get_keyword_condition}
\title{Make sub-clause for given keywords}
\usage{
get_keyword_condition(keyword_list)
}
\arguments{
\item{keyword_list}{a list of string keywords on which to filter the dataset}
}
\value{
filter sub-clause string
}
\description{
Make sub-clause for given keywords
}
\keyword{query}
