% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_polarity_keras.R
\name{get_predictions}
\alias{get_predictions}
\title{Get predictions for the given subset of data}
\usage{
get_predictions(data, maxlen, model_load_path, trained_data, max_words)
}
\arguments{
\item{data}{Given subset of data to make predictions on.}

\item{maxlen}{Maximum length of a sequence.}

\item{model_load_path}{File path location of the trained model file.}

\item{trained_data}{Training dataset.}

\item{max_words}{Maximum number of words to consider using word frequency measure.}
}
\value{
Given subset of data appended with pred_polarity column for predicted values.
}
\description{
Run given keras model to make predictions on the given subset of data
}
