% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{get_time_condition}
\alias{get_time_condition}
\title{Make sub-clause for given time range}
\usage{
get_time_condition(start_date_time, end_date_time)
}
\arguments{
\item{start_date_time}{input start_date_time in POSIXct format on which to filter the dataset}

\item{end_date_time}{input end_date_time in POSIXct format on which to filter the dataset}
}
\value{
filter sub-clause string
}
\description{
Make sub-clause for given time range
}
\keyword{query}
