% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\name{get_tokenizer}
\alias{get_tokenizer}
\title{Get Tokenizer for tokenizing train/test data}
\usage{
get_tokenizer(data, max_words)
}
\arguments{
\item{data}{Data to fit tokenizer on.}

\item{max_words}{Maximum number of words to consider using word frequency measure.}
}
\value{
tokenizer object
}
\description{
Get Tokenizer for tokenizing train/test data
}
