% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{get_user_list_condition}
\alias{get_user_list_condition}
\title{Make sub-clause for given users}
\usage{
get_user_list_condition(user_list)
}
\arguments{
\item{user_list}{a vector of users for which to filter the dataset.}
}
\value{
filter sub-clause string
}
\description{
Make sub-clause for given users
}
\keyword{query}
