% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{is_valid_sentiment_data_frame}
\alias{is_valid_sentiment_data_frame}
\title{check if a data frame is valid sentiment data set}
\usage{
is_valid_sentiment_data_frame(data)
}
\arguments{
\item{data}{data frame to check}
}
\value{
bool true or false
}
\description{
Method to check if a data frame is valid sentiment data set
}
