% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\docType{data}
\name{keras_config_params}
\alias{keras_config_params}
\title{Constants for keras model training}
\format{An object of class \code{list} of length 3.}
\usage{
keras_config_params
}
\description{
Constants for keras model training
}
\keyword{datasets}
