% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{make_clause}
\alias{make_clause}
\title{Make clause for given parameters}
\usage{
make_clause(user_list, start_date_time, end_date_time, keyword_list)
}
\arguments{
\item{user_list}{a vector of users for which to filter the dataset.}

\item{start_date_time}{input start_date_time in POSIXct format on which to filter the dataset}

\item{end_date_time}{input end_date_time in POSIXct format on which to filter the dataset}

\item{keyword_list}{a list of string keywords on which to filter the dataset}
}
\value{
filter clause string
}
\description{
Make clause for given parameters
}
\keyword{validation}
