% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_helpers.R
\name{parse_glove_embeddings}
\alias{parse_glove_embeddings}
\title{Parse glove embeddings}
\usage{
parse_glove_embeddings(file_path)
}
\arguments{
\item{file_path}{File path location of the downloaded glove embeddings.}
}
\value{
Parsed embeddings index object.
}
\description{
Parse glove embeddings
}
