% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_polarity_keras.R
\name{predict_polarity_keras}
\alias{predict_polarity_keras}
\title{Predict polarity for the given dataset.}
\arguments{
\item{data}{the sentiment140 test \code{text} for text of the tweet.}

\item{user_list}{a vector of users for which to filter the dataset}

\item{start_date_time}{input start_date_time in POSIXct format on which to filter the dataset}

\item{end_date_time}{input end_date_time in POSIXct format on which to filter the dataset}

\item{keyword_list}{a list of string keywords on which to filter the dataset}

\item{maxlen}{Maximum length of a sequence.}

\item{max_words}{Maximum number of words to consider using word frequency measure.}

\item{trained_data}{Training dataset.}

\item{model_load_path}{File path location of the trained model file.}
}
\value{
a list object with \code{raw} filtered dataframe, \code{predictions} dataframe that holds the predicted polarity using the model, \code{confusion_matrix}, multiple model performance statistics  and a \code{plot} comparing the actual and predicted polarity.
}
\description{
Predict polarity for the given dataset and filters using the pre-trained Keras LSTM model.
}
\examples{
library(deepSentimentR)
predict_polarity_keras()

}
\keyword{keras}
\keyword{modelling}
\keyword{visualization}
