% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment140_test.R
\docType{data}
\name{sentiment140_test}
\alias{sentiment140_test}
\title{\code{sentiment140_test} test dataset from Sentiment140 (\url{http://help.sentiment140.com/for-students/})}
\description{
\code{sentiment140_test} dataset used as test dataset for sentiment140 analysis.
}
\keyword{data}
