% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment140_train.R
\docType{data}
\name{sentiment140_train}
\alias{sentiment140_train}
\title{A \code{sentiment140_train} training dataset from Sentiment140}
\description{
A \code{sentiment140_train} dataset used as training dataset for sentiment140 analysis.
}
\keyword{data}
