% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{subsample_input_data}
\alias{subsample_input_data}
\title{subsample input data to max size}
\usage{
subsample_input_data(data, data_size = config_params$max_input_data_size)
}
\arguments{
\item{data}{input data}

\item{data_size}{defaults to the max input data size}
}
\value{
sub sampled data if needed.
}
\description{
subsample input data to max size
}
