% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{time_series}
\alias{time_series}
\title{Generate time series plots to analyze data based on filters}
\arguments{
\item{data}{the sentiment140 train or test data containing variables \code{user} for username, \code{date} for date and
\code{text} for text of the tweet.}

\item{user_list}{a vector of users for which to filter the dataset}

\item{start_date_time}{input start_date_time in POSIXct format on which to filter the dataset}

\item{end_date_time}{input end_date_time in POSIXct format on which to filter the dataset}

\item{keyword_list}{a list of string keywords on which to filter the dataset}
}
\value{
a list object with \code{raw} filtered dataframe, \code{date_counts} aggregated dataframe that holds the frequency counts of date by polarity, \code{day_counts} aggregated dataframe that holds the frequency counts of day by polarity and a \code{plots} depicting their relationship.
}
\description{
Generate time series plots to analyze given data with filters and visualize over polarity and return the raw data, date/day aggregate data and a \code{plots} for the same.
}
\examples{
data(sentiment140_train)
time_series(user_list = c("DarkPiano"))

}
\keyword{analysis}
\keyword{visualization}
