% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_conv_1d_with_glove.R
\name{train_conv1d_with_glove}
\alias{train_conv1d_with_glove}
\title{Train 1-dimensional Convolution Network with Glove Embeddings}
\arguments{
\item{data}{the sentiment140 train dataset with \code{text} for text of the tweet and \code{polarity} for polarity.}

\item{max_words}{Maximum number of words to consider using word frequency measure.}

\item{maxlen}{Maximum length of a sequence.}

\item{embedding_dim}{Output dimension of the embedding layer.}

\item{epochs}{Number of epochs to run the training for.}

\item{batch_size}{Batch Size for model fitting.}

\item{validation_split}{Split ratio for validation}

\item{conv1d_filters}{Number of filters i.e. output dimension for convolution layers.}

\item{conv1d_kernel_size}{Window size for convolution layers.}

\item{conv1d_pool_size}{Pool size for max pooling.}

\item{seed}{Seed for shuffling training data.}

\item{glove_file_path}{File path location for glove embeddings.}

\item{model_save_path}{File path location for saving model.}
}
\value{
plot of the training operation showing train vs validation loss and accuracy.
}
\description{
Train 1-dimensional Convolution Network using keras on the given dataset using Glove Embeddings available here - (\url{https://nlp.stanford.edu/projects/glove/})
}
\examples{
\dontrun{
  data(sentiment140_train)
  train_conv1d_with_glove(glove_file_path = "./glove.6B.100d.txt",
                          model_save_path = "./train_glove_conv_1d.h5")
}

}
\keyword{keras}
\keyword{modelling}
