% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_lstm_with_glove.R
\name{train_lstm_with_glove}
\alias{train_lstm_with_glove}
\title{Train LSTM model with Glove Embeddings}
\arguments{
\item{data}{the sentiment140 train dataset with \code{text} for text of the tweet and \code{polarity} for polarity.}

\item{max_words}{Maximum number of words to consider using word frequency measure.}

\item{maxlen}{Maximum length of a sequence.}

\item{embedding_dim}{Output dimension of the embedding layer.}

\item{epochs}{Number of epochs to run the training for.}

\item{batch_size}{Batch Size for model fitting.}

\item{validation_split}{Split ratio for validation}

\item{lstm_units}{Number of units i.e. output dimension of lstm layer.}

\item{seed}{Seed for shuffling training data.}

\item{glove_file_path}{File path location for glove embeddings.}

\item{model_save_path}{File path location for saving model.}
}
\value{
plot of the training operation showing train vs validation loss and accuracy.
}
\description{
Train LSTM model using keras on the given dataset using Glove Embeddings available here - (\url{https://nlp.stanford.edu/projects/glove/})
}
\examples{
\dontrun{
  data(sentiment140_train)
  train_lstm_with_glove(glove_file_path = "./glove.6B.100d.txt",
                        model_save_path = "./train_glove_lstm.h5")
}

}
\keyword{keras}
\keyword{modelling}
