% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{validate_list}
\alias{validate_list}
\title{Method to validate a list}
\usage{
validate_list(input_list, message)
}
\arguments{
\item{input_list}{input list}

\item{message}{message text to report in case of failure (optional)}
}
\value{
Execution halts with error if invalid format found.
}
\description{
Method to validate a list
}
