% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{validate_sentiment_data_frame}
\alias{validate_sentiment_data_frame}
\title{Method to validate a data frame}
\usage{
validate_sentiment_data_frame(data)
}
\arguments{
\item{data}{input data}
}
\value{
Execution halts with error if invalid format found.
}
\description{
Method to validate a data frame
}
