% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram_network.R
\name{visualize_bigrams}
\alias{visualize_bigrams}
\title{Generate network plot for bigram and counts}
\usage{
visualize_bigrams(bigrams)
}
\arguments{
\item{bigrams}{dataframe containing bigrams and their counts}
}
\value{
plot object with network visualization
}
\description{
Generate network plot for bigram and counts with arrows pointing to bigram word direction and edge alpha relative to counts
}
\keyword{bigrams}
