% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cor_network.R
\name{visualize_cor}
\alias{visualize_cor}
\title{Generate network plot for words and their pairwise correlation}
\usage{
visualize_cor(cors)
}
\arguments{
\item{cors}{dataframe containing words and their pairwise correlation}
}
\value{
plot object with network visualization
}
\description{
Generate network plot for words and their correlations with edge alpha relative to correlation.
}
\keyword{bigrams}
