% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_cor_network.R
\name{word_cor_network}
\alias{word_cor_network}
\title{Generate word correlation network}
\arguments{
\item{data}{the sentiment140 train or test data containing variables \code{user} for username, \code{date} for date and
\code{text} for text of the tweet.}

\item{user_list}{a vector of users for which to filter the dataset}

\item{start_date_time}{input start_date_time in POSIXct format on which to filter the dataset}

\item{end_date_time}{input end_date_time in POSIXct format on which to filter the dataset}

\item{keyword_list}{a list of string keywords on which to filter the dataset}

\item{correlation_threshold}{threshold beyond which to plot the network}
}
\value{
a list object with \code{raw} filtered dataframe, \code{word_cors} aggregated dataframe that holds the correlated words  and a \code{plot} representing the network.
}
\description{
Generate word correlation network to analyze given data with filters for all highly correlated words excluding stop words.
}
\examples{
data(sentiment140_test)
word_cor_network()

}
\keyword{analysis}
\keyword{visualization}
