% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment140_train.R
\docType{data}
\name{sentiment140_train}
\alias{sentiment140_train}
\title{\code{sentiment140_train} training dataset from Sentiment140 (\url{http://help.sentiment140.com/for-students/})}
\description{
\code{sentiment140_train} dataset used as training dataset for sentiment analysis.
}
\keyword{data}
